requirements.txt
Flask==2.3.2
python-barcode==0.14.0
Pillow==10.2.0
qrcode==7.4.2
